Make a Cipher OS Environment Variable

make_cipher_env() creates a Windows OS environment variable string from the provided cipher. OS environment variables beginning with 'remoter_' and ending with the target session string are searched when calling connect_remote with credentials = "ENV".

make_cipher_env(
  cipher = NULL,
  shared_key = NULL,
  session = make.names(tolower(Sys.getenv("COMPUTERNAME")))
)

Arguments

cipher

If NULL,the R image holding encryption objects is interactively chosen; otherwise, an object created with make_cipher.

shared_key

If the shared key was not included in the cipher object, this value is checked and throws an error if empty

session

The name of the session for which the environment variable string is created

Value

a JSON string that can be stored in a Windows OS environment variable formatted as "remoter_"