Manage a remoter Session

server_fun starts or stops a remoter session (see remoter-package for documentation).

server_fun(
  auth_root = path.expand("~"),
  server_dir = "~",
  session = make.names(tolower(Sys.getenv("COMPUTERNAME"))),
  ...
)

Arguments

auth_root

The path to the authentication objects to read ('.rdata' files) containing the ciphers and decryption keys. These should be generated from make_cipher

server_dir

The path to the working directory for the spawned server

session

(string) The session prefix for the ciphers loaded from auth_root

...

dots_list: additional arguments passed externally

Note

Designed for Windows OS