A remoter Wrapper

connect_remote is a wrapper for remoter::client making use of predefined encrypted authentication objects

connect_remote(
  credentials = NULL,
  prompt = "REMOTE_SESSION::",
  port = NULL,
  action = "client",
  session = make.names(tolower(Sys.getenv("COMPUTERNAME"))),
  ...
)

Arguments

credentials

One of the following:

string: The path to the R workspace image containing the cipher and key needed to connect to the remote resource

TRUE: Choose the R workspace image interactively via tk_choose.files

'ENV': See 'Details'

NULL|FALSE: Use when no credentials are needed (connections will not be secure)

prompt

(string) The prompt to use during the remote session

port

(integer) The port to use if not provided in the cipher (useful for specific overrides or ad-hoc sessions)

action

(string, symbol) The function to use to connect to the remote session (e.g., client, batch, etc.)

session

(string) The session prefix for the ciphers when credentials is 'ENV'

...

Not used

Details

When credentials is passed 'ENV', a Windows environment variable named 'remoter_' must already be set. The content of the variable should be a JSON string created with make_cipher_env. Expected keys are as follows:

addr: string (IPV4 format)

port: numeric,string

session: string

cipher: string (hexidecimal)

nonce: string (hexidecimal, must evaluate to length 24L raw vector)

shared_key: string (hexidecimal)

References

At a minimum, read the following documentation:

client

c2s

s2c

exit