A remoter Wrapper

connect_remote is an R6 class serving as a wrapper for remoter::client and remoter::batch making use of predefined encrypted authentication objects. Functionality is designed to work on Windows operating systems.

Note

No defaults other than addr, port, and password are provided.

References

At a minimum, read the following documentation:

client

c2s

s2c

exit

Public fields

prompt

A string to use as the prompt when connecting interactively to a remote session (defaults to "REMOTE_SESSION")

Active bindings

history

Returns an environment object from which the history of captured connection output can be accessed

Methods

Public methods

connect_remote$new()

connect_remote$print()

connect_remote$connect()

connect_remote$clone()

Method new()

When credentials is passed 'ENV', a Windows environment variable named 'remoter_' must already be set. The content of the variable should be a JSON string created with make_cipher_env. Expected keys are as follows:

addr: string (IPV4 format)

port: numeric,string

session: string

cipher: string (hexidecimal)

nonce: string (hexidecimal, must evaluate to length 24L raw vector)

shared_key: string (hexidecimal)

Usage

connect_remote$new(
  credentials = NULL,
  prompt = "REMOTE_SESSION",
  port = NULL,
  session = make.names(tolower(Sys.getenv("COMPUTERNAME"))),
  ...
)

Arguments

credentials

One of the following:

string: The path to the R workspace image containing the cipher and key needed to connect to the remote resource

TRUE: Choose the R workspace image interactively via tk_choose.files

'ENV': See 'Details'

NULL|FALSE: Use when no credentials are needed (connections will not be secure)

prompt

(string) The prompt to use during the remote session

port

(integer) The port to use if not provided in the cipher (useful for specific overrides or ad-hoc sessions)

session

(string) The session prefix for the ciphers when credentials is 'ENV'

...

Not used

Method print()

The print method

Usage

connect_remote$print()

Method connect()

The saved authentication objects are used on-demand to make the connection.

Usage

connect_remote$connect(action = "client", capture = FALSE, ...)

Arguments

action

(string, symbol) The remoter function to use to connect to the remote session: client (default) or batch.

capture

(logical) Should the remote output be captured?

...

Additional arguments to use.

Returns

The class environment invisibly

Method clone()

The objects of this class are cloneable with this method.

Usage

connect_remote$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.